Separate eval to its own module
