+++ eval/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rair-eval"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+rair-cmd = {workspace = true}
+rair-core = {workspace = true}
